Updates Terms of Service acceptance flow
Refactors the Terms of Service (ToS) acceptance flow to use a UUID stored in the settings instead of a boolean.

This enables better tracking of ToS updates and ensures users are prompted to accept new versions.

The ToS content is now loaded from a resource file.
Additional file system permissions have been added to allow reading the file.