Adds clipboard manager plugin to the application
Integrates the tauri clipboard manager plugin to enable clipboard access functionality.

Adds the clipboard-manager plugin as a dependency.
Updates the desktop capability configuration to include clipboard-manager permissions.
Removes unused react color picker library.

The default permission for the clipboard-manager plugin is set to be explicitly enabled, as we believe
the clipboard can be inherently dangerous and it is
application-specific if read and/or write access is needed.