Adds user ban handling and event emission
Implements user ban detection and handling for both Warframe Market and Quantframe APIs.

The changes introduce a mechanism to detect user bans based on API responses. When a ban is detected, the system emits an event to update the UI and stop analytics collection to prevent further API calls. This includes the addition of callback functions to handle different ban scenarios and update the UI accordingly. Also added the ability to disconnect the user from the API's.

The goal is to improve user experience by informing users about their ban status and preventing the app from making further requests, as well as improving detection and handling of user ban scenarios.