Adds user permission check functionality
Implements a permission flag system for users.

This allows to restrict access to certain features based on user permissions.
A new `PermissionsFlags` enum is introduced to define available permissions.
The `User` struct now has a `has_permission` method to check if a user has a specific permission.
An AuthGuard has been implemented, will be added later, to protect the routes.