Adds item panel and export permission check
Adds the Item panel to the live scraper and integrates it into the Tauri client.

Implements permission checks for data export functionality in the Riven, Item and WishList panels.
This ensures that users can only export data if they have the necessary permissions.
The permission check is performed on mount and the export button is disabled accordingly.

Refactors `HasPermission` function to return a Promise for asynchronous handling.