Refetches data on transaction refresh event
Listens for the `RefreshTransactions` Tauri event and triggers a data refetch in both the home page summary and transaction panel.
This ensures that the UI reflects the latest transaction data after an update.
Removes unnecessary dependency from the `useTauriEvent` hook in transaction panel.
Updates the width of the right section in transaction panel.