Adds Riven ranking and Remoraid link
Adds the Riven ranking feature and integrates a link to Remoraid.dev, providing users with a tool to create new designs.

Also fixes an issue where buying directly from the WFM tab with a quantity greater than 1 would miscalculate the price and fixes an issue where user was able to type in the chat box.

Addresses inconsistencies in the release notes by reordering and correcting the placement of certain features.