Adds minimum price to wish list items
Adds a minimum price field to wish list items.

This change allows users to set a minimum price for items in their wish list.
The scraper will then use this price to update the list price if the scraper price is less than the minimum.

Also makes temporary database used for debugging purposes